+++ src/index.ts
@@ -9,9 +9,11 @@ import { ticketReply } from './commands/builders/ticketReply';
 import { addRole } from './commands/builders/addRole';
 import { removeRole } from './commands/builders/removeRole';
 import { getRoles } from './commands/builders/getRoles';
+import lang from './utils/lang.json';
+import { botSetup } from './commands/builders/botSetup';
+import { setStatus } from './utils/setStatus';
 dotenv.config();
 
-const GUILD = process.env.GUILD_ID!;    // (want to eventually deprecate this)
 const RELEASE = process.env.RELEASE!;   // determines which bot we're using
 let TOKEN = process.env.BOT_TOKEN!;     // default production bot token
 let CLIENT = process.env.CLIENT_ID!;    // default production bot client
@@ -21,7 +23,7 @@ if (RELEASE == 'dev') {
     TOKEN = process.env.DEV_BOT_TOKEN!;
     CLIENT = process.env.DEV_CLIENT_ID!;
     // log that we're using the development bot
-    console.log('');
+    console.log(lang.general.usingDev);
 }
 
 const client = new Client({
@@ -37,6 +39,7 @@ const rest = new REST({ version: '10' }).setToken(TOKEN);
 
 /* Slash Commands */
 const commands: RESTPostAPIApplicationCommandsJSONBody[] = [
+    botSetup,       // bot setup
     addRole,        // add a role
     removeRole,     // remove a role
     getRoles,       // get roles
@@ -44,13 +47,9 @@ const commands: RESTPostAPIApplicationCommandsJSONBody[] = [
     ticketReply,    // ticket reply
 ];
 
-client.once('ready', () => {
-    console.log(`Logged in as ${client.user?.tag}`);
-});
-
 client.on('interactionCreate', async (interaction) => {
     if (interaction.isChatInputCommand()) {
-      client.emit('chatInputCommand', client, interaction);
+        client.emit('chatInputCommand', client, interaction);
     } else if (interaction.isButton() || interaction.isStringSelectMenu() || interaction.isModalSubmit()) {
         client.emit('buttonInteraction', client, interaction);
     }
@@ -61,19 +60,24 @@ client.on('chatInputCommand', handleSlashCommand);
 // when button is pressed, handleTicketInteraction will run
 client.on('buttonInteraction', handleTicketInteraction);
 
+// once we reday, LEGGO
+client.once('ready', () => {
+    // log that we logged in
+    console.log(`Logged in as ${client.user?.tag}`);
+
+    // set status
+    setStatus(client);
+});
+
 // main function to do all the things
 async function main() {
     try {
-        // remove any 'guild specific' commands we had in place
-        // (this will eventually be removed once i'm done doing the things)
-        await rest.put(Routes.applicationGuildCommands(CLIENT, GUILD), { 
-                body: [], // empty array removes all commands
-            });
 
+        /*
         // remove any application commands we had in place (refresh it basically)
         await rest.put(Routes.applicationCommands(CLIENT), {
             body: [],
-        });
+        });*/
 
         // register le commands we currently have
         await rest.put(Routes.applicationCommands(CLIENT), {
@@ -82,6 +86,8 @@ async function main() {
 
         // initialize typeORM shtuff
         await AppDataSource.initialize();
+
+        // log in
         client.login(TOKEN);
     } catch(err) {
         console.log("Startup Error:", err);