+++ src/typeorm/index.ts
@@ -6,6 +6,7 @@ import { ArchivedTicketConfig } from './entities/ArchivedTicketConfig';
 import { ArchivedTicket } from './entities/ArchivedTicket';
 import { ServerConfig } from './entities/ServerConfig';
 import { SavedRole } from './entities/SavedRole';
+import { BotConfig } from './entities/BotConfig';
 dotenv.config();
 
 export const AppDataSource = new DataSource({
@@ -16,5 +17,5 @@ export const AppDataSource = new DataSource({
     password: process.env.MYSQL_DB_PASSWORD,
     database: process.env.MYSQL_DB_DATABASE,
     synchronize: true,
-    entities: [TicketConfig, Ticket, ArchivedTicketConfig, ArchivedTicket, ServerConfig, SavedRole],
+    entities: [TicketConfig, Ticket, ArchivedTicketConfig, ArchivedTicket, ServerConfig, SavedRole, BotConfig],
 });